move all curve tests to a single test file
